update the docker image
* update the docker image repository

The owner moved the container image to a new repository.

* use return to finalize error execution when retryCount is 0

This lets the flow of control fully exit the request and not continue
on when not desired, which should let the process fully finish. This
particularly shows up when running unit tests, and Jest is unable to
complete.

* update docker command argument unit tests

* tweak docker argument unit test to validate all arguments

expect.arrayContaining() only checks a subset of the target array, so we
were missing checks for the new arguments.

* update README with new docker image